Moved Routes to new file
Moved to new file for modularization and better readability
