+++ data.json
@@ -1,14 +1,1 @@
-{
-  "items": [
-    { "name": "Item 2", "description": "This is item 2" },
-    { "name": "Item 2" },
-    { "name": "Item 2", "description": "item 2 description", "price": 3434 },
-    { "name": "Item 2", "description": "item 2 description", "price": 3434 },
-    {
-      "name": "Item",
-      "description": "this is item 2",
-      "price": 32,
-      "stock": 30
-    }
-  ]
-}
+{"items":[{"name":"Item 2","description":"This is item 2"},{"name":"Item 2"},{"name":"Item 2","description":"item 2 description","price":3434},{"name":"Item 2","description":"item 2 description","price":3434},{"name":"Item","description":"this is item 2","price":32,"stock":30},{"name":"Item","description":"this is item 2","price":89,"stock":30},{"name":"Item 6","description":"this is item 5","price":89,"stock":30}]}
